Set up Map
I made the main map, each box correspond to a stand.
I created empty sub levels for each Level/Minigame.
Please be sure to make different branches for each level you will work
on and to double click on YOUR level before starting work on it.
Also hide the other levels (by clicking on the eye you see on the left
hand side in the level menu).
Remember that every time you save the project on UE4, the changes will
appear on GitHub as well (click on Changes on the top menu).
If you have any question please write them on Basecamp